Enhances parser to support PEP723 inline script metdata
Enhances parse_dependencies to recognize dependencies which are
declared using the "Inline script metadata" specification. This was
originally developed in PEP-723 (https://peps.python.org/pep-0723/)
and is now managed by the Python Packaging Authority, where it
is described in the PyPA specification here:

https://packaging.python.org/en/latest/specifications/inline-script-metadata/#inline-script-metadata

This specification defines a syntax which different tools may adopt to
read metadata embedded in a python script. A particular piece of
metadata identifies the meaning of a piece of metadata by an
ASCII identifier of its "TYPE".

This commit enhances pythonrunscript so that it recognizes the
following TYPE identifiers:

pythonrunscript-requirements-txt
pythonrunscript-environment-yml
pythonrunscript-conda-install-specs-txt

For instance, with this new syntax, one can embed a requirements.txt
file as follows:

> # /// pythonrunscript-requirements-txt
> # requests==2.26.0
> # """
> # ///